+++ src/native_fs.rs
@@ -6,7 +6,10 @@ use std::str;
 /// the root filesystem API
 pub struct FileSystem {}
 /// synchronous (native) / asynchronous (web) file API
-pub struct File(std::fs::File);
+pub enum File {
+    Native(std::fs::File),
+}
+
 pub type IoError = std::io::Error;
 pub type IoErrorKind = std::io::ErrorKind;
 
@@ -17,9 +20,21 @@ impl FileSystem {
     /// You can then open this file with `FileSystem::open("config.json")`.
     /// When running your native project, use `cd www && cargo run`.
     /// When deploying on the web, the file should simply be in the same directory as index.html, as config.json.
+    /// Note that on wasm target, this works with any URL, you can do :
+    /// `FileSystem::open('https://raw.githubusercontent.com/unrust/uni-app/master/www/test.txt')`
+    /// To be able to load data from an URL in native mode, you have to enable the http feature.
     pub fn open(s: &str) -> Result<File, IoError> {
-        let file = std::fs::File::open(s)?;
-        Ok(File(file))
+        if s.starts_with("http") {
+            Err(std::io::Error::new(
+                ErrorKind::Other,
+                format!(
+                    "error : could not read {} : http feature not enabled in uni-app::FileSystem",
+                    s
+                ),
+            ))
+        } else {
+            Ok(File::Native(std::fs::File::open(s)?))
+        }
     }
 }
 
@@ -27,15 +42,21 @@ impl File {
     /// Once the file has been loaded (see [`File::is_ready`]), returns the file content as `Vec<u8>`
     pub fn read_binary(&mut self) -> Result<Vec<u8>, IoError> {
         let mut buf = Vec::new();
-        self.0.read_to_end(&mut buf)?;
+        match self {
+            File::Native(f) => {
+                f.read_to_end(&mut buf)?;
+            }
+        }
         Ok(buf)
     }
     /// Once the file has been loaded (see [`File::is_ready`]), returns the file content as a String
     pub fn read_text(&mut self) -> Result<String, IoError> {
         let mut data = String::new();
-        match self.0.read_to_string(&mut data) {
-            Ok(_) => Ok(data),
-            Err(e) => Err(std::io::Error::new(ErrorKind::Other, e)),
+        match self {
+            File::Native(f) => match f.read_to_string(&mut data) {
+                Ok(_) => Ok(data),
+                Err(e) => Err(std::io::Error::new(ErrorKind::Other, e)),
+            },
         }
     }
     /// return true if the file has been loaded